+++ tests/test_diagram_errors.py
@@ -10,6 +10,7 @@ from cemento.draw_io.constants import (
     BlankTermLabelError,
     DisconnectedTermError,
     FloatingEdgeError,
+    InvertedEdgeError,
     MissingChildEdgeError,
     MissingParentEdgeError,
 )
@@ -110,7 +111,7 @@ def test_no_labels_disconnection_errors():
 def test_all_arrow_types():
     errors = get_diagram_errors(input_path=diagram_test_files[5])
     expected_error_types = {
-        BidirectionalEdgeError: 4,
+        BidirectionalEdgeError: 2,
         FloatingEdgeError: 5,
     }
     check_errors_by_count(errors, expected_error_types)
@@ -123,6 +124,18 @@ def test_bidirectional_arrow_errors():
         MissingParentEdgeError: 1,
         MissingChildEdgeError: 1,
         BidirectionalEdgeError: 4,
+        FloatingEdgeError: 1,
+    }
+    check_errors_by_count(errors, expected_error_types)
+
+
+def test_inverted_arrow_errors():
+    errors = get_diagram_errors(input_path=diagram_test_files[7])
+    expected_error_types = {
+        DisconnectedTermError: 6,
+        MissingParentEdgeError: 1,
+        MissingChildEdgeError: 1,
+        InvertedEdgeError: 4,
         FloatingEdgeError: 1,
     }
     check_errors_by_count(errors, expected_error_types)